# openapi.yml
openapi: 3.0.2
info:
  description: Moov ACH ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating ACH files. ACH is the primary method of electronic money movement throughout the United States.
  version: v1
  title: ACH API
  contact:
    url: https://github.com/$(OURLY)/ach
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /files:
    get:
      tags: ['ACH Files']
      summary: Gets a list of Files
      operationId: getFiles
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
      responses:
        '200':
          description: A list of File objects
          headers:
            X-Total-Count:
              description: The total number of Originators
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Files'
  /files/create:
    post:
      tags: ['ACH Files']
      summary: Create a new File object
      operationId: createFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: a4f88150
          required: false
          schema:
            type: string
      requestBody:
        description: Content of the ACH file (in json or raw text)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFile'
          text/plain:
            schema:
              description: A plaintext ACH file
              type: string
              example: 101 222380104 1210428821805100000A094101Citadel                Bank Name
      responses:
        '201':
          description: A JSON object containing a new File
          headers:
            Location:
              description: The location of the new resource
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          description: "Invalid File Header Object"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /files/{fileID}:
    get:
      tags: ['ACH Files']
      summary: Retrieves the details of an existing File. You need only supply the unique File identifier that was returned upon creation.
      operationId: getFileByID
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      responses:
        '200':
          description: A File object for the supplied ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '404':
          description: A resource with the specified ID was not found
    post:
      tags: ['ACH Files']
      summary: Updates the specified File Header by setting the values of the parameters passed. Any parameters not provided will be left unchanged.
      operationId: updateFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: a4f88150
          required: false
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFile'
      responses:
        '201':
          description: A JSON object containing a new File
          headers:
            Location:
              description: The location of the new resource
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          description: "Invalid File Header Object"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: ['ACH Files']
      summary: Permanently deletes a File and associated Batches. It cannot be undone.
      operationId: deleteACHFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
      responses:
          '200':
            description: Permanently deleted File.
          '404':
            description: A File with the specified ID was not found.
  /files/{fileID}/contents:
    get:
      tags: ['ACH Files']
      summary: Assembles the existing file (batches and controls) records, computes sequence numbers and totals. Returns plaintext file.
      operationId: getFileContents
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      responses:
        '200':
          description: File built successfully without errors.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RawFile'
  /files/{fileID}/validate:
    get:
      tags: ['ACH Files']
      summary: Validates the existing file. You need only supply the unique File identifier that was returned upon creation.
      operationId: validateFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      responses:
        '200':
          description: File validated successfully without errors.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          description: Validation failed. Check response for errors
  /files/{fileID}/segment:
    post:
      tags: ['ACH Files']
      summary: Create a new file object
      operationId: segmentFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: a4f88150
          required: false
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      requestBody:
        description: Content of the ACH file (in json or raw text)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFile'
          text/plain:
            schema:
              description: A plaintext ACH file
              type: string
              example: 101 222380104 1210428821805100000A094101Citadel                Bank Name
      responses:
        '201':
          description: A JSON object containing the credit File ID and debit File ID
          headers:
            Location:
              description: The location of the new resource
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
  /files/{fileID}/flatten:
    post:
      tags: ['ACH Files']
      summary: Create a new file object
      operationId: flattenFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: a4f88150
          required: false
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      requestBody:
        description: Content of the ACH file (in json or raw text)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFile'
          text/plain:
            schema:
              description: A plaintext ACH file
              type: string
              example: 101 222380104 1210428821805100000A094101Citadel                Bank Name
      responses:
        '201':
          description: A JSOn objeect containing the File ID associated with a Flattened ACH file.
          headers:
            Location:
              description: The location of the new resource
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
  /files/{fileID}/batches:
    get:
      tags: ['ACH Files']
      summary: Get the batches on a File.
      operationId: getFileBatches
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      responses:
        '200':
          description: A list of Batch objects
          headers:
            X-Total-Count:
              description: The total number of Batches on the File.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batches'
    post:
      tags: ['ACH Files']
      summary: Add Batch to File
      operationId: addBatchToFile
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Idempotent key in the header which expires after 24 hours. These strings should contain enough entropy for to not collide with each other in your requests.
          example: a4f88150
          required: false
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Batch'
      responses:
        '200':
          description: Batch added to File
  /files/{fileID}/batches/{batchID}:
    get:
      tags: ['ACH Files']
      summary: Get a specific Batch on a FIle
      operationId: getFileBatch
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
        - name: batchID
          in: path
          description: Batch ID
          required: true
          schema:
            type: string
            example: 45758063
      responses:
        '200':
          description: Batch object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '404':
          description: Batch or File not found
    delete:
      tags: ['ACH Files']
      summary: Delete a Batch from a File
      operationId: deleteFileBatch
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional Request ID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: fileID
          in: path
          description: File ID
          required: true
          schema:
            type: string
            example: 3f2d23ee214
        - name: batchID
          in: path
          description: Batch ID
          required: true
          schema:
            type: string
            example: 45758063
      responses:
        '200':
          description: Batch deleted
        '404':
          description: Batch or File not found
components:
  schemas:
    CreateFile:
      properties:
        ID:
          type: string
          description: File ID
          example: 3f2d23ee214
        fileHeader:
          $ref: '#/components/schemas/FileHeader'
        batches:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
        IATBatches:
          type: array
          items:
            $ref: '#/components/schemas/IATBatch'
        fileControl:
          $ref: '#/components/schemas/FileControl'
      required:
        - fileHeader
    File:
      properties:
        ID:
          type: string
          description: File ID
          example: 3f2d23ee214
        fileHeader:
          $ref: '#/components/schemas/FileHeader'
        batches:
          type: array
          items:
            $ref: '#/components/schemas/Batch'
        IATBatches:
          type: array
          items:
            $ref: '#/components/schemas/IATBatch'
        fileControl:
          $ref: '#/components/schemas/FileControl'
        # NotificationOfChange: # TODO(adam)
        # ReturnEntries: # TODO(adam)
    FileHeader:
      properties:
        immediateOrigin:
          type: string
          description: contains the Routing Number of the ACH Operator or sending point that is sending the file.
          minLength: 9
          maxLength: 10
          example: "99991234"
        immediateOriginName:
          type: string
          description: The name of the ACH operator or sending point that is sending the file.
          maxLength: 23
          example: My Bank Name
        immediateDestination:
          type: string
          maxLength: 10
          minLength: 9
          example: "69100013"
          description: contains the Routing Number of the ACH Operator or receiving point to which the file is being sent
        immediateDestinationName:
          type: string
          description: The name of the ACH or receiving point for which that file is destined.
          maxLength: 23
          example: Federal Reserve Bank
        fileCreationTime:
          description: 'The File Creation Date is the date when the file was prepared by an ODFI. (Format HHmm - H=Hour, m=Minute)'
          type: string
          example: 1504
        fileCreationDate:
          description: 'The File Creation Time is the time when the file was prepared by an ODFI. (Format YYMMDD - Y=Year, M=Month, D=Day)'
          type: string
          example: 190102
        fileIDModifier:
          type: string
          description: Incremented value for each file for RDFI's.
          example: 0
      required:
        - immediateOrigin
        - immediateOriginName
        - immediateDestination
        - immediateDestinationName
    FileControl:
      properties:
        ID:
          description: Moov API File ID
          type: string
          example: d1e26288
        batchCount:
          description: Count of Batches in the File
          type: integer
          example: 1
        blockCount:
          description: |
            BlockCount total number of records in the file (include all headers and trailer) divided by 10 (This number must be evenly divisible by 10. If not, additional records consisting of all 9’s are added to the file after the initial ‘9’ record to fill out the block 10.)
          type: integer
          example: 1
        entryAddendaCount:
          description: Total detail and addenda records in the file
          type: integer
          example: 0
        entryHash:
          description: EntryHash calculated in the same manner as the batch has total but includes total from entire file
          type: integer
          example: 0
        totalDebit:
          description: Accumulated Batch debit totals within the file.
          type: integer
          example: 100
        totalCredit:
          description: Accumulated Batch credit totals within the file.
          type: integer
          example: 20
    RawFile:
      type: string
      description: Plaintext ACH file
      example: "101 222380104 1210428821805100000A094101Citadel                Bank Name"
    Files:
      type: array
      items:
        $ref: '#/components/schemas/File'
    Batch:
      properties:
        batchHeader:
          $ref: '#/components/schemas/BatchHeader'
        entryDetails:
          type: array
          items:
            $ref: '#/components/schemas/EntryDetail'
        batchControl:
          $ref: '#/components/schemas/BatchControl'
    BatchHeader:
      required:
        - serviceClassCode
        - companyName
        - companyIdentification
        - ODFIIdentification
      properties:
        ID:
          type: string
          description: Batch Header ID
          example: 913b5742
        serviceClassCode:
          type: integer
          description: Service Class Code - ACH Credits Only 220 and ACH Debits Only 225
          example: 220
        companyName:
          type: string
          description: Company originating the entries in the batch
          example: Acme Corp
        companyDiscretionaryData:
          type: string
          description: The 9 digit FEIN number (proceeded by a predetermined alpha or numeric character) of the entity in the company name field
          example: 123456789
        standardEntryClassCode:
          type: string
          description: Identifies the payment type (product) found within an ACH batch-using a 3-character code.
          example: PPD
        companyEntryDescription:
          type: string
          description: |
            A description of the entries contained in the batch.
            The Originator establishes the value of this field to provide a description of the purpose of the entry to be displayed back to the receive For example, "GAS BILL," "REG. SALARY," "INS. PREM,", "SOC. SEC.," "DTC," "TRADE PAY," "PURCHASE," etc.
            This field must contain the word "REVERSAL" (left justified) when the batch contains reversing entries.
            This field must contain the word "RECLAIM" (left justified) when the batch contains reclamation entries.
            This field must contain the word "NONSETTLED" (left justified) when the batch contains entries which could not settle.
          example: PURCHASE
        companyDescriptiveDate:
          type: string
          description: |
            The Originator establishes this field as the date it would like to see displayed to the receiver for descriptive purposes. This field is never used to control timing of any computer or manual operation. It is solely for descriptive purposes. The RDFI should not assume any specific format.
        effectiveEntryDate:
          description: 'Date on which the entries are to settle. Format YYMMDD (Y=Year, M=Month, D=Day)'
          type: string
          example: 190102
        originatorStatusCode:
          type: integer
          description: |
            ODFI initiating the Entry.
            0 ADV File prepared by an ACH Operator.
            1 This code identifies the Originator as a depository financial institution.
            2 This code identifies the Originator as a Federal Government entity or agency.
        ODFIIdentification:
          description: First 8 digits of the originating DFI transit routing number
          type: string
          example: 12345678
        batchNumber:
          type: string
          description: |
            BatchNumber is assigned in ascending sequence to each batch by the ODFI or its Sending Point in a given file of entries. Since the batch number in the Batch Header Record and the Batch Control Record is the same, the ascending sequence number should be assigned by batch and not by record.
...
    Batches:
      type: array
      items:
        $ref: '#/components/schemas/Batch'
...
    IATBatch:
      properties:
        ID:
          description: Client defined string used as a reference to this record.
          type: string
          example: a747e53f
        IATBatchHeader:
          $ref: '#/components/schemas/IATBatchHeader'
...
    Error:
      required:
        - error
      properties:
        error:
          type: string
          description: An error message describing the problem intended for humans.
          example: Validation error(s) present.